requirements.txt
flash-attn
litgpt>=0.5.7
torch>=2.5.0
lightning>=2.5.0
transformers>=4.47.0
datasets>=3.1.0
tokenizers>=0.21.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.19.0
tqdm>=4.66.0
torchmetrics>=1.0.0
pyyaml>=6.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.6.0
sentencepiece>=0.2.0
safetensors>=0.4.0
peft>=0.13.0
einops>=0.8.0
accelerate>=0.34.0
bitsandbytes>=0.44.0
deepspeed>=0.15.0
tensorboard>=2.0.0
matplotlib>=3.8.0
rich>=11.2.0